+++ packages/gov-ui-kit/src/core/components/avatars/avatarBase/avatarBase.tsx
@@ -1,0 +1,12 @@
+import { forwardRef, type ComponentPropsWithoutRef } from 'react';
+import { useOdsCoreContext } from '../../odsCoreProvider';
+
+export interface IAvatarBaseProps extends ComponentPropsWithoutRef<'img'> {}
+
+export const AvatarBase = forwardRef<HTMLImageElement, IAvatarBaseProps>((props, ref) => {
+    const { Img } = useOdsCoreContext();
+
+    return <Img ref={ref} {...props} />;
+});
+
+AvatarBase.displayName = 'AvatarBase';