+++ packages/gov-ui-kit/src/core/components/link/link/index.ts
@@ -1,0 +1,2 @@
+export { Link } from './link';
+export { type ILinkProps, type LinkVariant } from './link.api';