+++ packages/gov-ui-kit/src/core/components/link/link/link.api.ts
@@ -1,5 +1,5 @@
 import type { AnchorHTMLAttributes } from 'react';
-import { type IconType } from '../icon';
+import { type IconType } from '../../icon';
 
 export type LinkVariant = 'primary' | 'neutral';