+++ packages/gov-ui-kit/src/core/components/link/link/link.tsx
@@ -1,9 +1,10 @@
 import classNames from 'classnames';
 import React from 'react';
-import { Icon } from '../icon';
+import { Icon } from '../../icon';
+import { LinkBase } from '../linkBase';
 import type { ILinkProps, LinkVariant } from './link.api';
 
-const variantToLabelClassNames: Record<LinkVariant, string[]> = {
+const variantToLabelClassNames: Record<LinkVariant | 'disabled', string[]> = {
     primary: [
         'text-primary-400 cursor-pointer', // Default
         'hover:text-primary-600', // Hover state
@@ -14,9 +15,8 @@ const variantToLabelClassNames: Record<LinkVariant, string[]> = {
         'hover:text-neutral-800', // Hover state
         'active:text-neutral-800', // Active state
     ],
+    disabled: ['truncate text-neutral-300'],
 };
-
-const disabledStyle = 'truncate text-neutral-300';
 
 export const Link = React.forwardRef<HTMLAnchorElement, ILinkProps>((props, ref) => {
     const {
@@ -33,34 +33,39 @@ export const Link = React.forwardRef<HTMLAnchorElement, ILinkProps>((props, ref)
         ...otherProps
     } = props;
 
+    const processedVariant = disabled ? 'disabled' : variant;
     const linkClassName = classNames(
         'inline-flex max-w-fit flex-col gap-y-0.5 truncate rounded text-sm leading-tight',
         'focus:outline-none focus-visible:ring focus-visible:ring-primary focus-visible:ring-offset',
         'md:gap-y-1 md:text-base',
-        disabled ? disabledStyle : variantToLabelClassNames[variant],
+        variantToLabelClassNames[processedVariant],
         className,
     );
 
-    const descriptionClassName = classNames('truncate', disabled ? disabledStyle : 'text-neutral-500');
     const linkRel = target === '_blank' ? `noopener noreferrer ${rel}` : rel;
 
     return (
-        <a
+        <LinkBase
             ref={ref}
             onClick={!disabled ? onClick : undefined}
             href={!disabled ? href : undefined}
             className={linkClassName}
             target={target}
             rel={linkRel}
-            {...(disabled && { tabIndex: -1, 'aria-disabled': 'true' })}
+            tabIndex={disabled ? -1 : undefined}
+            aria-disabled={disabled}
             {...otherProps}
         >
             <div className="flex items-center gap-x-2">
                 <span className="truncate">{children}</span>
                 {iconRight && <Icon icon={iconRight} size="sm" />}
             </div>
-            {description && <p className={descriptionClassName}>{description}</p>}
-        </a>
+            {description && (
+                <p className={classNames('truncate', disabled ? 'text-neutral-300' : 'text-neutral-500')}>
+                    {description}
+                </p>
+            )}
+        </LinkBase>
     );
 });